Add Parser that can parser classes
So far the Parser can only parse class names. I want separate parsing
into different classes, but I couldn't think of a good way to do it.

One idea I had was to pass in the lexer to a class and call .next(), as
I did in the switch statement.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>