Omit parens if there are no parameters
Methods can omit the parens if there are no parameters.

    init() {} -> init {}

Because of the way we're parsing, fields can be declared with no type.

    class Test {

        empty

        init {
        }

    }

Will be parsed as:

    Class: Test
      Field: [] void empty
      Method: [public] Test init ([])

I don't think this should be included because this can lead to unwanted
bugs. If you accidently type `a` in the middle of your code, everything
is parsed wrong.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>