Modularize Parser
"Top" parsing, class parsing, and method parsing have been extracted
out of Parser and put into separate classes.

"Top" represents the outside of a class, aka the "file" that is being
parsed. "Files" are not limited to one class, and they're not really
limited to being in a file. Any string of text can be passed into the
Parser, and the top of the tree is parsed by TopParser.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>